Preliminary code for gui config changes.
Base code added to allow instrument configurations to be changed using
the gui instead of modifying the .json files manually. Instruments
must provide a class that looks like an instrument.ConfigWindow.

A DataWindow class has also been added to instrument.py solely for the
purpose of documenting the required methods for a DataWindow (thanks
to python's duck typing).