Allowed a subset of the DC channels to be selected
In the configuration for the DC power supply it is
now possible to select a subset of the channels
available. Channels are also now selected by name
rather than number.

The rationale for this change was to remove any
unused channels from the results when using any
channel after the first.